+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/scripting/actions/ActionFactory.kt
@@ -37,6 +37,7 @@ import ch.rmy.android.http_shortcuts.scripting.actions.types.SendTCPPacketAction
 import ch.rmy.android.http_shortcuts.scripting.actions.types.SendUDPPacketActionType
 import ch.rmy.android.http_shortcuts.scripting.actions.types.SetResultActionType
 import ch.rmy.android.http_shortcuts.scripting.actions.types.SetVariableActionType
+import ch.rmy.android.http_shortcuts.scripting.actions.types.SetWireguardTunnelStateActionType
 import ch.rmy.android.http_shortcuts.scripting.actions.types.ShareTextActionType
 import ch.rmy.android.http_shortcuts.scripting.actions.types.TextToSpeechActionType
 import ch.rmy.android.http_shortcuts.scripting.actions.types.ToHexStringActionType
@@ -91,6 +92,7 @@ constructor(
     sendUDPPacketActionType: SendUDPPacketActionType,
     setResultActionType: SetResultActionType,
     setVariableActionType: SetVariableActionType,
+    setWireguardTunnelStateActionType: SetWireguardTunnelStateActionType,
     shareTextActionType: ShareTextActionType,
     textToSpeechActionType: TextToSpeechActionType,
     toastActionType: ToastActionType,
@@ -150,6 +152,7 @@ constructor(
             sendUDPPacketActionType,
             setResultActionType,
             setVariableActionType,
+            setWireguardTunnelStateActionType,
             shareTextActionType,
             textToSpeechActionType,
             toastActionType,